Refactored code and added utils for login and fetching data.
