+++ LegoCollectionManager/LegoCollectionManager/Views/User/Create.cshtml
@@ -23,6 +23,7 @@
                 <input type="submit" value="Register" class="btn btn-primary" />
             </div>
         </form>
+        @ViewBag.InvalidUser.ToString()
     </div>
 </div>
